% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clean_port_data}
\alias{clean_port_data}
\title{load and clean port data}
\usage{
clean_port_data(data)
}
\arguments{
\item{data}{data loaded usikng load_port_data and wait to be cleaned}
}
\description{
load and clean port data
}
