% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{fit_arima_forecast}
\alias{fit_arima_forecast}
\title{Fit arima forecast model}
\usage{
fit_arima_forecast(target, exog)
}
\arguments{
\item{target}{a n*1 vector, output of armia}

\item{exog}{a n*m matrix, exogenous variables of armia}
}
\description{
Fit arima forecast model
}
