% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{analyze_arima_result}
\alias{analyze_arima_result}
\title{Fit arima forecast model}
\usage{
analyze_arima_result(result)
}
\arguments{
\item{result}{result to be analyzed}
}
\description{
Fit arima forecast model
}
